Expand update_image for services to support init containers.
This includes adding test coverage for existing behavior to
preserve backwards-compatibility.

This was written to be a minimal addition to the existing
support for 'additional containers', which work essentially the
same as 'init containers'. To keep the changes minimal, unit
tests rely on mocks and patches instead of rewriting the code
to be more easily unit testable.