Patch config file into test_at_least_host.
This ensures it doesn't depend on a specific
local config being present, which means it
will pass out of box for new users.

This relies on a mock instead of setting the
config_file arg of the client to a YAML in the
files/ directory partly so it doesn't have to
make any changes to the test itself to make it
pass and partly so the test continues to exercise
the client's default behavior.